+++ src/Attributes/Skip.php
@@ -4,17 +4,26 @@ declare(strict_types=1);
 namespace MyTester\Attributes;
 
 use Attribute;
+use MyTester\ISkipAttribute;
 
 /**
  * Skip attribute
- * Defines conditions when a test should be skipped
+ * Allows skipping a test unconditionally
  *
  * @author Jakub Konečný
  */
 #[Attribute(Attribute::TARGET_METHOD)]
-final class Skip
+final readonly class Skip implements ISkipAttribute
 {
-    public function __construct(public array $value = [])
+    public string $value;
+
+    public function __construct()
     {
+        $this->value = "";
+    }
+
+    public function getSkipValue(): string
+    {
+        return $this->value;
     }
 }