+++ src/SkipChecker.php
@@ -19,13 +19,8 @@ final class SkipChecker implements ISkipChecker
 {
     public const string ANNOTATION_NAME = "skip";
 
-    /**
-     * @var callable[]
-     * @deprecated
-     */
-    private array $checkers = [];
-
     private array $skipAnnotations = [
+        "skip" => Attributes\Skip::class,
         "requiresPhpVersion" => Attributes\RequiresPhpVersion::class,
         "requiresPhpExtension" => Attributes\RequiresPhpExtension::class,
         "requiresSapi" => Attributes\RequiresSapi::class,
@@ -34,29 +29,6 @@ final class SkipChecker implements ISkipChecker
 
     public function __construct(private readonly Reader $annotationsReader)
     {
-        $this->addChecker("php", [$this, "checkPhpVersion"]); // @phpstan-ignore method.deprecated
-        $this->addChecker("extension", [$this, "checkLoadedExtension"]); // @phpstan-ignore method.deprecated
-        $this->addChecker("sapi", [$this, "checkPhpSapi"]); // @phpstan-ignore method.deprecated
-        $this->addChecker("osFamily", [$this, "checkOsFamily"]); // @phpstan-ignore method.deprecated
-    }
-
-    /**
-     * @deprecated Use new attributes instead
-     */
-    public function addChecker(string $name, callable $callback): void
-    {
-        $this->checkers[$name] = $callback;
-    }
-
-    /**
-     * @param class-string $class
-     * @deprecated
-     */
-    public function getSkipValue(string $class, string $method): ?array
-    {
-        /** @var array|null $value */
-        $value = $this->annotationsReader->getAnnotation(self::ANNOTATION_NAME, $class, $method);
-        return $value;
     }
 
     /**
@@ -71,77 +43,13 @@ final class SkipChecker implements ISkipChecker
                 $attribute = new $classname($value);
                 $skipValue = $attribute->getSkipValue();
                 if (is_string($skipValue)) {
+                    if ($skipValue === "") {
+                        return true;
+                    }
                     return $skipValue;
                 }
             }
         }
-
-        $value = $this->getSkipValue($class, $method); // @phpstan-ignore method.deprecated
-        if ($value === null) {
-            return false;
-        }
-        if ($value === []) {
-            return true;
-        }
-        foreach ($value as $k => $v) {
-            $checker = Arrays::get($this->checkers, $k, null); // @phpstan-ignore property.deprecated
-            if ($checker === null) {
-                return false;
-            }
-            $value = $checker($v);
-            if (is_string($value)) {
-                return $value;
-            }
-        }
         return false;
-    }
-
-    /**
-     * @deprecated Use attribute {@see RequiresPhpVersion} instead
-     */
-    public function checkPhpVersion(mixed $value): ?string
-    {
-        if (!is_string($value) && !is_numeric($value)) {
-            return null;
-        }
-        trigger_error("Using an array to skip a test is deprecated, use an attribute instead", E_USER_DEPRECATED);
-        return (new Attributes\RequiresPhpVersion((string) $value))->getSkipValue();
-    }
-
-    /**
-     * @deprecated Use attribute {@see RequiresPhpExtension} instead
-     */
-    public function checkLoadedExtension(mixed $value): ?string
-    {
-        if (!is_string($value)) {
-            return null;
-        }
-        trigger_error("Using an array to skip a test is deprecated, use an attribute instead", E_USER_DEPRECATED);
-        return (new Attributes\RequiresPhpExtension($value))->getSkipValue();
-    }
-
-    /**
-     * @deprecated Use attribute {@see RequiresSapi} instead
-     */
-    public function checkPhpSapi(mixed $value): ?string
-    {
-        if (!is_string($value)) {
-            return null;
-        }
-        trigger_error("Using an array to skip a test is deprecated, use an attribute instead", E_USER_DEPRECATED);
-        return (new Attributes\RequiresSapi($value))->getSkipValue();
-    }
-
-    /**
-     * @see PHP_OS_FAMILY
-     * @deprecated Use attribute {@see RequiresOsFamily} instead
-     */
-    public function checkOsFamily(mixed $value): ?string
-    {
-        if (!is_string($value)) {
-            return null;
-        }
-        trigger_error("Using an array to skip a test is deprecated, use an attribute instead", E_USER_DEPRECATED);
-        return (new Attributes\RequiresOsFamily($value))->getSkipValue();
     }
 }